requirements.txt
google-cloud-storage
requests
folium
